Made loop_start/end similar to other instructions.
By moving logic into the loop, at the cost of violating recipe a bit more,
a uniform interface is available for instruction creation.
Without default getattr we avoid ambiguities between an
instruction which is not implemented and one which is actually illegal. This used to cause fake loops to be created
when an instruction wasn't understood.

It also avoids ambiguities due to similarity of the 'the' form with loops and other instructions.