Added multiple recipe parsing support.
Each Recipe is now packed in a RecipeNode in the AST.
All parse functions pass around an Environment object which
holds a dictionary with all recipes for now.

Accessing preserve.recipes as before would result in other
modules finding recipes always empty since that is its parse
time value, and in Python, that will be copied into the including module's environment.